Assimp initial implementation and some refactors

+++ PhoenixEngine/dep/assimp/types.h
@@ -358,7 +358,7 @@ struct aiString {
         if (!len) {
             return;
         }
-        if (length + len >= MAXLEN) {
+        if (length + static_cast<unsigned long long>(len) >= MAXLEN) {
             return;
         }
